Few Small changes
Did away with the CSS contiainers they will get their own class if i
picke it up again.
Some small changes. I plan to.....
Renamed A file to make name patterns const.
Added a ImageFor to the Mix. Will accept the image as a byte[] or as a
bitmap for now. Writes it as a <img> using a base 64 data string. Tis
was simply load from your database :).
Rewrote the FileFor will now compile vs leverage input for. Allows for
classes to be used to add attributes once more.